requirements.txt
fastapi
uvicorn
llama-cpp-python
pyautogen
argparse
colorama
